Add min-rate parameter
This parameter is actually a syntactic sugar for a specific combination
of connections , timeout and concurency so scylla-bench will send `min-rate`
requests per second to scylla. It is important to note that this can (
as in the real usecase) to cause some wrong results in the statistics.
The idea of this patch is to create a combination of those parameters
that will effectively send requests to the server but will give up on
a lot of the result in order not to bloat memory.